requirements.txt
Django==4.2.1
djangorestframework==3.14.0
Markdown==3.4.3
django_filter==23.2
black==23.3.0
mypy==1.3.0
flake8==6.0.0
isort==5.12.0
pre-commit==3.3.2
psycopg2-binary==2.9.6
django-environ==0.10.0
drf-yasg==1.21.5
pytest==7.3.1
pytest-django==4.5.2
model-bakery==1.11.0
django-filter==23.2
